sched/idle: Reported an error when an illegal negative value is passed
hulk inclusion
category: bugfix
bugzilla: 180842 https://gitee.com/openeuler/kernel/issues/I4DDEL

Reported an error when an illegal negative value is passed. The allowed value is
0 to INT_MAX.

+++ include/linux/kernel.h
@@ -556,7 +556,7 @@ extern int sysctl_panic_on_stackoverflow;
 
 extern bool crash_kexec_post_notifiers;
 #ifdef CONFIG_IAS_SMART_IDLE
-extern unsigned long poll_threshold_ns;
+extern int poll_threshold_ns;
 #endif
